refactor: update import paths for utility functions in UI components
