+++ libs/shared/src/components/ui/accordion.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 import * as AccordionPrimitive from "@radix-ui/react-accordion"
-import { cn } from "./utils"
+import { cn } from "../../lib/utils"
 import { ChevronDownIcon } from "@radix-ui/react-icons"
 
 const Accordion = AccordionPrimitive.Root
