+++ libs/shared/src/components/ui/context-menu.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 import * as ContextMenuPrimitive from "@radix-ui/react-context-menu"
-import { cn } from "./utils"
+import { cn } from "../../lib/utils"
 import { CheckIcon, ChevronRightIcon, DotFilledIcon } from "@radix-ui/react-icons"
 
 const ContextMenu = ContextMenuPrimitive.Root
