+++ libs/shared/src/components/ui/hover-card.tsx
@@ -3,7 +3,7 @@
 import * as React from "react"
 import * as HoverCardPrimitive from "@radix-ui/react-hover-card"
 
-import { cn } from "./utils"
+import { cn } from "../../lib/utils"
 
 const HoverCard = HoverCardPrimitive.Root
