+++ libs/shared/src/components/ui/input-otp.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 import { OTPInput, OTPInputContext } from "input-otp"
-import { cn } from "./utils"
+import { cn } from "../../lib/utils"
 import { MinusIcon } from "@radix-ui/react-icons"
 
 const InputOTP = React.forwardRef<
