+++ libs/shared/src/components/ui/navigation-menu.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import * as NavigationMenuPrimitive from "@radix-ui/react-navigation-menu"
 import { cva } from "class-variance-authority"
-import { cn } from "./utils"
+import { cn } from "../../lib/utils"
 import { ChevronDownIcon } from "@radix-ui/react-icons"
 
 const NavigationMenu = React.forwardRef<
