+++ libs/shared/src/components/ui/resizable.tsx
@@ -1,7 +1,7 @@
 "use client"
 import * as ResizablePrimitive from "react-resizable-panels"
 
-import { cn } from "./utils"
+import { cn } from "../../lib/utils"
 import { DragHandleDots2Icon } from "@radix-ui/react-icons"
 
 const ResizablePanelGroup = ({
