+++ libs/shared/src/components/ui/skeleton.tsx
@@ -1,4 +1,4 @@
-import { cn } from "./utils"
+import { cn } from "../../lib/utils"
 
 function Skeleton({
   className,
