+++ libs/shared/src/components/ui/textarea.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { cn } from "./utils"
+import { cn } from "../../lib/utils"
 
 const Textarea = React.forwardRef<
   HTMLTextAreaElement,
